+++ CryptopalsChallenge/ByteEncryption.h
@@ -22,7 +22,7 @@ public:
 	~ByteEncryption();
 
 	// pass false to encrypt to perform decryption.
-	static void aes_ecb_encrypt_block(byte *input, byte *key, int keyLengthBytes, byte *output, bool encrypt);
+	static void aes_ecb_encrypt_block(byte *input, byte *key, size_t keyLengthBytes, byte *output, bool encrypt);
 	static void aes_ecb_encrypt(ByteVector *bv, ByteVector *key, ByteVector *output, size_t start_index, size_t end_index, bool encrypt);
 	static void aes_cbc_encrypt(ByteVector *bv, ByteVector *key, ByteVector *output, ByteVector *iv, bool encrypt);
